+++ helpers/formatStdErr.js
