+++ routes/api.js
@@ -5,11 +5,11 @@ const fs = require('fs-extra');
 const FormData = require('form-data');
 const multer = require("multer");
 var router = express.Router();
-const transcribe = require('../transcribe-api-wrapped')
-const transcribeWrapped = require("../transcribe-wrapped");
-const constants = require("../constants");
+const transcribe = require('../transcribe/transcribe-api-wrapped')
+const transcribeWrapped = require("../transcribe/transcribe-wrapped");
+const constants = require("../constants/constants");
 const filenamify = require("filenamify");
-const createTranslatedFiles = require("../translate-files-api");
+const createTranslatedFiles = require("../translate/translate-files-api");
 const { languagesToTranslateTo, newLanguagesMap, translationLanguages } = constants;
 
 const makeFileNameSafe = function(string){
