+++ routes/index.js
@@ -4,19 +4,19 @@ const multer = require("multer");
 var router = express.Router();
 const FormData = require('form-data');
 const fs = require('fs-extra');
-const downloadAndTranscribe = require('../download.js')
+const downloadAndTranscribe = require('../downloading/download.js')
 // const transcribe = require('../transcribe');
-const transcribeWrapped = require('../transcribe-wrapped');
+const transcribeWrapped = require('../transcribe/transcribe-wrapped');
 const Queue = require("promise-queue");
-const {languagesToTranscribe} = require("../constants");
+const {languagesToTranscribe} = require("../constants/constants");
 const filenamify = require("filenamify");
-const { forHumans, forHumansNoSeconds } = require('../helpers')
+const { forHumans, forHumansNoSeconds } = require('../helpers/helpers')
 const path = require('path');
 const moment = require('moment');
 const ffprobe = require("ffprobe");
 const which = require("which");
 const {fr} = require("language-name-map/map");
-const { languagesToTranslateTo, newLanguagesMap, modelsArray, whisperLanguagesHumanReadableArray } = require('../constants');
+const { languagesToTranslateTo, newLanguagesMap, modelsArray, whisperLanguagesHumanReadableArray } = require('../constants/constants');
 const ffprobePath = which.sync('ffprobe')
 const _ = require('lodash');
